+++ package-dev/Runtime/SentryInitialization.cs
@@ -3,6 +3,8 @@
 #define SENTRY_NATIVE_IOS
 #elif UNITY_ANDROID
 #define SENTRY_NATIVE_ANDROID
+#elif UNITY_STANDALONE_WIN && ENABLE_IL2CPP
+#define SENTRY_NATIVE_WINDOWS
 #endif
 #endif
 
@@ -13,6 +15,8 @@ using UnityEngine.Scripting;
 using Sentry.Unity.iOS;
 #elif UNITY_ANDROID
 using Sentry.Unity.Android;
+#elif SENTRY_NATIVE_WINDOWS
+using Sentry.Unity.Native;
 #endif
 
 [assembly: AlwaysLinkAssembly]
@@ -27,11 +31,14 @@ namespace Sentry.Unity
             var options = ScriptableSentryUnityOptions.LoadSentryUnityOptions();
             if (options.ShouldInitializeSdk())
             {
+                var sentryUnityInfo = new SentryUnityInfo();
 
 #if SENTRY_NATIVE_IOS
                 SentryNativeIos.Configure(options);
 #elif SENTRY_NATIVE_ANDROID
-                SentryNativeAndroid.Configure(options);
+                SentryNativeAndroid.Configure(options, sentryUnityInfo);
+#elif SENTRY_NATIVE_WINDOWS
+                SentryNative.Configure(options);
 #endif
 
                 SentryUnity.Init(options);