+++ samples/unity-of-bugs/Assets/Scripts/MobileNativeSupport/NativeSupportButtons.cs
@@ -69,7 +69,7 @@ public class NativeSupportButtons : MonoBehaviour
 #if ENABLE_IL2CPP
         throw_cpp();
 #else
-        Debug.Log("Requires IL2CPP. Try this on a native player.");
+        Debug.Log("Requires IL2CPP. Try this on a native player that supports C++ native plugins.");
 #endif
     }
 
@@ -78,7 +78,7 @@ public class NativeSupportButtons : MonoBehaviour
 #if ENABLE_IL2CPP
         crash_in_cpp();
 #else
-        Debug.Log("Requires IL2CPP. Try this on a native player.");
+        Debug.Log("Requires IL2CPP. Try this on a native player that supports C++ native plugins.");
 #endif
     }
 
@@ -87,7 +87,7 @@ public class NativeSupportButtons : MonoBehaviour
 #if ENABLE_IL2CPP
         crash_in_c();
 #else
-        Debug.Log("Requires IL2CPP. Try this on a native player.");
+        Debug.Log("Requires IL2CPP. Try this on a native player that supports C native plugins.");
 #endif
     }
 
@@ -108,7 +108,7 @@ public class NativeSupportButtons : MonoBehaviour
 #if PLATFORM_IOS
         throwObjectiveC();
 #else
-        Debug.Log("Requires IL2CPP. Try this on a native player.");
+        Debug.Log("Requires IL2CPP. Try this on a native player that supports Objective-C native plugins.");
 #endif
     }