+++ test/Scripts.Tests/test-pack-contents.ps1
@@ -12,25 +12,25 @@ $packageFile = "$projectRoot/package-release.zip"
 Push-Location $projectRoot
 
 if (-not(Test-Path -Path $packageFile)) {
-    Write-Host  "Package '$packageFile' not found. 
+    Write-Host  "Package '$packageFile' not found.
 Run 'scripts/pack.ps1' first."
     exit 1
 }
 
 if (-not(Test-Path -Path $snapshotFile)) {
-    Write-Host  "Snapshot file '$snapshotFile' not found. 
+    Write-Host  "Snapshot file '$snapshotFile' not found.
 Can't compare package contents against baseline."
     exit 2
 }
 
 $zip = [IO.Compression.ZipFile]::OpenRead($packageFile)
 try {
-    $snapshotContent = $zip.Entries
+    $snapshotContent = $zip.Entries.FullName.Replace("\", "/")
     if ($args.Count -gt 0 -and $args[0] -eq "accept") {
         # Override the snapshot file with the current package contents
-        $snapshotContent.FullName | Out-File $snapshotFile
+        $snapshotContent | Out-File $snapshotFile
     }
-    $result = Compare-Object $snapshotContent.FullName (Get-Content $snapshotFile)
+    $result = Compare-Object $snapshotContent (Get-Content $snapshotFile)
     Write-Host  $result
     if ($result.count -eq 0)
     {