+++ test/Sentry.Unity.Tests/IntegrationTests.cs
@@ -189,6 +189,11 @@ namespace Sentry.Unity.Tests
         {
             yield return SetupSceneCoroutine("1_BugFarm");
 
+            // We should use the sample Dsn for the nextDsn
+            // to avoid static dsn.
+            var options = AssetDatabase.LoadAssetAtPath(ScriptableSentryUnityOptions.GetConfigPath(ScriptableSentryUnityOptions.ConfigName),
+                typeof(ScriptableSentryUnityOptions)) as ScriptableSentryUnityOptions;
+
             var sourceEventCapture = new TestEventCapture();
             var sourceDsn = "https://94677106febe46b88b9b9ae5efd18a00@o447951.ingest.sentry.io/5439417";
             using var firstDisposable = InitSentrySdk(o =>
@@ -198,18 +203,40 @@ namespace Sentry.Unity.Tests
             });
 
             var nextEventCapture = new TestEventCapture();
-            var nextDsn = "https://a520c186ed684a8aa7d5d334bd7dab52@o447951.ingest.sentry.io/5801250";
+            var nextDsn = options?.Dsn;
             using var secondDisposable = InitSentrySdk(o =>
             {
                 o.Dsn = nextDsn;
                 o.AddIntegration(new UnityApplicationLoggingIntegration(eventCapture: nextEventCapture));
             });
-
             var testBehaviour = new GameObject("TestHolder").AddComponent<TestMonoBehaviour>();
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.TestException));
 
+            Assert.NotNull(nextDsn);
             Assert.AreEqual(0, sourceEventCapture.Events.Count, sourceDsn);
-            Assert.AreEqual(1, nextEventCapture.Events.Count, nextDsn);
+            Assert.AreEqual(1, nextEventCapture.Events.Count);
+        }
+
+        [UnityTest]
+        public IEnumerator BugFarmScene_DebugLogException_IsMarkedUnhandled()
+        {
+            yield return SetupSceneCoroutine("1_BugFarm");
+
+            // arrange
+            var testEventCapture = new TestEventCapture();
+            using var _ = InitSentrySdk(o =>
+            {
+                o.AddIntegration(new UnityApplicationLoggingIntegration(eventCapture: testEventCapture));
+            });
+            var testBehaviour = new GameObject("TestHolder").AddComponent<TestMonoBehaviour>();
+
+            // act
+            testBehaviour.gameObject.SendMessage(nameof(testBehaviour.DebugLogException));
+
+            // assert
+            Assert.NotNull(testEventCapture.Events.SingleOrDefault(sentryEvent =>
+                sentryEvent.SentryExceptions.SingleOrDefault(exception =>
+                    exception.Mechanism?.Handled is false) is not null));
         }
 
         [UnityTest]